Refactoring: Moved several classes and methods to new module 'openapc_toolkit' for reuse
